lib/stac-api/runtime/requirements.txt
mangum==0.15.1
stac-fastapi.api==2.4.1
stac-fastapi.extensions==2.4.1
stac-fastapi.pgstac==2.4.1
stac-fastapi.types==2.4.1
# https://github.com/stac-utils/stac-fastapi/pull/466
pygeoif==0.7
starlette_cramjam